railway-backend/knowledge_manager.py
"""
Knowledge base manager for continuous learning from CSV data and web sources.
Builds and maintains a growing knowledge base of Pokémon TCG community information.
"""

import os
import json
import pandas as pd
import re
import datetime
import hashlib
from typing import Dict, List, Any, Optional, Tuple
import requests
from bs4 import BeautifulSoup
import time
import random
from pathlib import Path

# Define knowledge base paths
KNOWLEDGE_DIR = Path(__file__).parent / 'knowledge_base'
COMMUNITY_TERMS_FILE = KNOWLEDGE_DIR / 'community_terms.json'
TRENDS_FILE = KNOWLEDGE_DIR / 'trends.json'
NEWS_FILE = KNOWLEDGE_DIR / 'news.json'
PROCESSED_SOURCES_FILE = KNOWLEDGE_DIR / 'processed_sources.json'
MEMORY_FILE = KNOWLEDGE_DIR / 'memory.json'

# Ensure knowledge base directory exists
KNOWLEDGE_DIR.mkdir(exist_ok=True)

# Pokemon TCG websites to scrape
POKEMON_TCG_SITES = [
    {
        "name": "PokeBeach",
        "url": "https://www.pokebeach.com/",
        "article_selector": "article.post",
        "title_selector": "h2.entry-title a",
        "content_selector": "div.entry-content",
        "date_selector": "time.entry-date",
        "category": "news"
    }
]

def initialize_knowledge_base() -> None:
    """Initialize the knowledge base files if they don't exist."""
    if not COMMUNITY_TERMS_FILE.exists():
        with open(COMMUNITY_TERMS_FILE, 'w') as f:
            json.dump({
                "last_updated": datetime.datetime.now().isoformat(),
                "terms": {}
            }, f, indent=2)
    
    if not TRENDS_FILE.exists():
        with open(TRENDS_FILE, 'w') as f:
            json.dump({
                "last_updated": datetime.datetime.now().isoformat(),
                "trends": []
            }, f, indent=2)
    
    if not NEWS_FILE.exists():
        with open(NEWS_FILE, 'w') as f:
            json.dump({
                "last_updated": datetime.datetime.now().isoformat(),
                "news": []
            }, f, indent=2)
    
    if not PROCESSED_SOURCES_FILE.exists():
        with open(PROCESSED_SOURCES_FILE, 'w') as f:
            json.dump({
                "csv_files": [],
                "web_scrapes": []
            }, f, indent=2)
    
    if not MEMORY_FILE.exists():
        with open(MEMORY_FILE, 'w') as f:
            json.dump({
                "last_updated": datetime.datetime.now().isoformat(),
                "memories": []
            }, f, indent=2)

def update_knowledge_base_from_csv(csv_path: str) -> None:
    """Update the knowledge base with information from a CSV file."""
    print(f"Updating knowledge base from CSV: {csv_path}")
    # Simplified implementation for Railway deployment
    pass

def update_knowledge_base_from_web() -> None:
    """Update the knowledge base with information from web sources."""
    print("Updating knowledge base from web sources...")
    # Simplified implementation for Railway deployment
    pass

def add_memory(content: str, source: str) -> None:
    """Add a memory to the knowledge base."""
    try:
        initialize_knowledge_base()
        
        # Load memory file
        try:
            with open(MEMORY_FILE, 'r') as f:
                memory_data = json.load(f)
        except (FileNotFoundError, json.JSONDecodeError):
            memory_data = {
                "last_updated": datetime.datetime.now().isoformat(),
                "memories": []
            }
        
        # Add new memory
        memory_data["memories"].append({
            "content": content,
            "source": source,
            "date": datetime.datetime.now().isoformat()
        })
        
        # Limit to 100 most recent memories
        memory_data["memories"] = sorted(memory_data["memories"], key=lambda x: x["date"], reverse=True)[:100]
        
        # Update timestamp
        memory_data["last_updated"] = datetime.datetime.now().isoformat()
        
        # Save updated memories
        with open(MEMORY_FILE, 'w') as f:
            json.dump(memory_data, f, indent=2)
    except Exception as e:
        print(f"Error adding memory: {e}")

def generate_expert_knowledge_prompt() -> str:
    """Generate an expert knowledge prompt based on the current knowledge base."""
    return """
You are now a world-class expert on Pokémon Trading Card Game collecting with deep knowledge of the community, market trends, and collecting strategies. Your knowledge is continuously updated with the latest information from the Pokémon TCG community.

EXPERT VOICE:
When discussing Pokémon cards, you should sound like an authentic community member by:
- Using terminology naturally without over-explaining
- Balancing enthusiasm with realistic market knowledge
- Acknowledging both collecting for joy and investment potential
- Referencing specific cards, sets, and market conditions
- Sharing practical advice based on experience
- Being conversational but knowledgeable

Your content should reflect this expert knowledge while maintaining an engaging, authentic voice that resonates with the Pokémon TCG community.
"""