railway-backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
tweepy>=4.14.0
python-dotenv>=1.0.0
pandas>=2.0.0
schedule>=1.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
openai>=1.0.0